add collection file

+++ src/controllers/vehicleControllers.js
@@ -2,6 +2,7 @@ import prisma from '../config/db.js';
 import { validationResult } from 'express-validator';
 
 export const vehicleController = {
+  // Add a vehicle
   addVehicle: async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -22,7 +23,6 @@ export const vehicleController = {
       dailyRate,
       status,
     } = req.body;
-    // const user_id = req.user.user_id;
 
     try {
       const vehicle = await prisma.vehicle.create({
@@ -39,20 +39,16 @@ export const vehicleController = {
           airConditioning,
           dailyRate,
           status,
-          // user_id,
         },
       });
       return res.status(201).json(vehicle);
     } catch (error) {
-      return res
-        .status(500)
-        .json({
-          error: 'Erreur lors de l\'ajout du véhicule: ' + error.message,
-        });
+      console.error('Add Vehicle Error:', error.message);
+      return res.status(500).json({ error: `Erreur lors de l'ajout du véhicule: ${error.message}` });
     }
   },
 
-  // Mettre à jour un véhicule
+  // Update a vehicle
   updateVehicle: async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -60,46 +56,31 @@ export const vehicleController = {
     }
 
     const { id } = req.params;
-    const dataToUpdate = {};
-
-    // Préparer les données à mettre à jour
-    for (const key in req.body) {
-      if (req.body[key] !== undefined) {
-        dataToUpdate[key] = req.body[key];
-      }
-    }
 
     try {
       const vehicle = await prisma.vehicle.update({
         where: { id: parseInt(id, 10) },
-        data: dataToUpdate,
+        data: req.body,
       });
       return res.status(200).json(vehicle);
     } catch (error) {
-      return res
-        .status(500)
-        .json({
-          error: 'Erreur lors de la mise à jour du véhicule: ' + error.message,
-        });
+      console.error('Update Vehicle Error:', error.message);
+      return res.status(500).json({ error: `Erreur lors de la mise à jour du véhicule: ${error.message}` });
     }
   },
 
-  // Récupérer tous les véhicules
+  // Retrieve all vehicles
   getAllVehicles: async (req, res) => {
     try {
       const vehicles = await prisma.vehicle.findMany();
       return res.status(200).json(vehicles);
     } catch (error) {
-      return res
-        .status(500)
-        .json({
-          error:
-            'Erreur lors de la récupération des véhicules: ' + error.message,
-        });
+      console.error('Get All Vehicles Error:', error.message);
+      return res.status(500).json({ error: `Erreur lors de la récupération des véhicules: ${error.message}` });
     }
   },
 
-  // Récupérer un véhicule par ID
+  // Retrieve a vehicle by ID
   getVehicleById: async (req, res) => {
     const { id } = req.params;
     try {
@@ -111,30 +92,54 @@ export const vehicleController = {
       }
       return res.status(200).json(vehicle);
     } catch (error) {
-      return res
-        .status(500)
-        .json({
-          error: 'Erreur lors de la récupération du véhicule: ' + error.message,
-        });
+      console.error('Get Vehicle By ID Error:', error.message);
+      return res.status(500).json({ error: `Erreur lors de la récupération du véhicule: ${error.message}` });
     }
   },
 
-  // Supprimer un véhicule
+  // Delete a vehicle
   deleteVehicle: async (req, res) => {
     const { id } = req.params;
+
     try {
+      // Vérifier s'il existe des contrats pour ce véhicule
+      const existingContracts = await prisma.contract.findMany({
+        where: {
+          vehicle_id: parseInt(id, 10),
+        },
+      });
+
+      if (existingContracts.length > 0) {
+        return res.status(400).json({
+          error: 'Impossible de supprimer le véhicule car il est lié à un contrat de location existant.',
+        });
+      }
+
+      // Vérifier s'il existe des réservations confirmées pour ce véhicule
+      const existingReservations = await prisma.reservation.findMany({
+        where: {
+          vehicle_id: parseInt(id, 10),
+          status: 'confirmed',
+        },
+      });
+
+      if (existingReservations.length > 0) {
+        return res.status(400).json({
+          error: 'Impossible de supprimer le véhicule car il a des réservations confirmées.',
+        });
+      }
+
+      // Supprimer le véhicule
       await prisma.vehicle.delete({
         where: { id: parseInt(id, 10) },
       });
-      return res
-        .status(204)
-        .json({ message: 'Véhicule supprimé avec succès.' });
+
+      return res.status(204).json({ message: 'Véhicule supprimé avec succès.' });
     } catch (error) {
-      return res
-        .status(500)
-        .json({
-          error: 'Erreur lors de la suppression du véhicule: ' + error.message,
-        });
+      console.error('Delete Vehicle Error:', error.message);
+      return res.status(500).json({ error: `Erreur lors de la suppression du véhicule: ${error.message}` });
+    } finally {
+      await prisma.$disconnect();
     }
   },
 };
